When only restarting, the entrypoint should be recreate the files

+++ deploy/bootstrap/lega-entrypoint.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 set -e
 
-cp /etc/ega/ssl.key /etc/ega/ssl.key.lega
-chmod 400 /etc/ega/ssl.key.lega
+if ! test -e /etc/ega/ssl.key.lega; then
+    cp /etc/ega/ssl.key /etc/ega/ssl.key.lega
+    chmod 400 /etc/ega/ssl.key.lega
+fi
 
 exec $@
